+++ src/main/java/coupon/data/CouponRepository.java
@@ -1,8 +1,6 @@
-package coupon;
+package coupon.data;
 
-import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CouponRepository extends JpaRepository<Coupon, Long> {
-    List<Coupon> findAll();
 }
